Trivial change
I was going to have :shell use the Git for Windows BASH emulator (sh.exe), but it turns out this is a bad idea becuase some plugins will assume that win32 environments have the :shell set to cmd.exe. Anyway, this lousy change is all that's left of my fiof my fi.